+++ src/opi/output/grepper/recipes.py
@@ -10,6 +10,7 @@ from opi.output.grepper.patterns import (
     HAS_FREQ,
     HAS_GEOMETRY_OPT,
     HAS_SCF,
+    NEB_CONVERGED,
     SCF_CONVERGED,
     TERMINATED_NORMALLY,
 )
@@ -259,6 +260,23 @@ def has_geometry_optimization_converged(file_name: Path, /) -> bool:
         True if expression is found in file else False
     """
     return has_string_in_file(file_name, GEOMETRY_CONVERGED)
+
+
+def has_neb_converged(file_name: Path, /) -> bool:
+    """
+    Searches for the message 'H U R R A Y' to indicate that a NEB optimization has converged.
+
+    Parameter
+    ---------
+    file_name: Path
+        Name of the output file
+
+    Returns
+    -------
+    bool
+        True if expression is found in file else False
+    """
+    return has_string_in_file(file_name, NEB_CONVERGED)
 
 
 def has_scf(file_name: Path, /) -> bool: